Backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
openai==1.30.0
pydantic==1.10.9
python-multipart==0.0.6
PyMuPDF==1.22.5      # fitz
python-docx==0.8.11
sentence-transformers==2.2.2
docx2txt==0.8
# numpy==1.26.0
numpy>=1.23.0

langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pypdf2
python-docx